requirements.txt
args==0.1.0
asn1crypto==0.22.0
astroid==1.5.3
bcrypt==3.1.3
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
clint==0.5.1
cloudinary==1.8.0
cookies==2.2.1
coverage==4.4.1
coveralls==1.1
cryptography==2.0.3
dj-database-url==0.4.2
Django==1.11.3
django-dotenv==1.4.1
django-redis==4.8.0
django-rq==0.9.5
docopt==0.6.2
future==0.16.0
gunicorn==19.7.1
httmock==1.2.6
httplib2==0.10.3
idna==2.5
isort==4.2.15
Jinja2==2.9.6
lazy-object-proxy==1.3.1
line-bot-sdk==1.3.0
linecache2==1.0.0
magicmock==0.3
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
natto==0.1.7
natto-py==0.8.0
numpy==1.13.1
olefile==0.44
pbr==3.1.1
Pillow==4.2.1
psycopg2==2.7.1
py==1.4.34
pycparser==2.18
PyJWT==1.5.2
pylint==1.7.1
pytest==3.0.7
pytest-cov==2.5.0
pytz==2017.2
redis==2.10.5
requests==2.18.1
requests-mock==1.3.0
requests-toolbelt==0.8.0
responses==0.5.1
rq==0.8.0
simple-salesforce==0.73.0
six==1.10.0
traceback2==1.4.0
trafaret==0.9.0
unittest2==1.1.0
urllib3==1.21.1
whitenoise==3.3.0
wrapt==1.10.10
